requirements.txt
flask
flask_restplus
flask-migrate
flask-sqlalchemy
sqlalchemy-utils
python-dotenv
